mock: Create new package and move mock code to it.
