+++ core/src/lib.rs
@@ -6,7 +6,6 @@
 #![allow(clippy::too_many_arguments)]
 
 use bitcoin::{OutPoint, Txid};
-// use clementine_circuits::{HashType, PreimageType};
 use serde::{Deserialize, Serialize};
 
 pub mod actor;
@@ -21,7 +20,6 @@ pub mod errors;
 pub mod extended_rpc;
 pub mod hashes;
 pub mod merkle;
-pub mod mock;
 pub mod musig2;
 pub mod operator;
 pub mod servers;
